mini-rag-chatbot/requirements.txt
# requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
transformers
torch
faiss-cpu
langdetect
pydantic
